+++ include/avtp/acf/CanBriefV2.h
@@ -341,8 +341,8 @@ OPEN1722_INLINE void Avtp_CanBriefV2_SetCanIdentifier(Avtp_CanBriefV2_t* msg, ui
  */
 OPEN1722_INLINE void Avtp_CanBriefV2_SetPayloadLen(Avtp_CanBriefV2_t* msg, uint16_t payloadLen) {
     uint16_t msgLenBytes = AVTP_CAN_BRIEF_V2_HEADER_LEN + payloadLen;
-    uint8_t pad = (4 - (msgLenBytes % 4)) % 4;
-    uint16_t msgLenQuadlets = (msgLenBytes + pad) / 4;
+    uint8_t pad = (uint8_t)(4 - (msgLenBytes % 4)) % 4;
+    uint16_t msgLenQuadlets = (uint16_t) ((msgLenBytes + pad) / 4);
     __Avtp_CanBriefV2_SetField(AVTP_CAN_BRIEF_V2_FIELD_ACF_MSG_LENGTH, msgLenQuadlets);
     __Avtp_CanBriefV2_SetField(AVTP_CAN_BRIEF_V2_FIELD_PAD, pad);
 }
