+++ include/avtp/acf/Gbb.h
@@ -453,8 +453,8 @@ OPEN1722_INLINE void Avtp_Gbb_SetSegmentNum(Avtp_Gbb_t* msg, uint16_t segmentNum
  */
 OPEN1722_INLINE void Avtp_Gbb_SetPayloadLen(Avtp_Gbb_t* msg, uint16_t payloadLen) {
     uint16_t msgLenBytes = AVTP_GBB_HEADER_LEN + payloadLen;
-    uint8_t pad = (4 - (msgLenBytes % 4)) % 4;
-    uint16_t msgLenQuadlets = (msgLenBytes + pad) / 4;
+    uint8_t pad = (uint8_t) (4 - (msgLenBytes % 4)) % 4;
+    uint16_t msgLenQuadlets = (uint16_t) ((msgLenBytes + pad) / 4);
     __Avtp_Gbb_SetField(AVTP_GBB_FIELD_ACF_MSG_LENGTH, msgLenQuadlets);
     __Avtp_Gbb_SetField(AVTP_GBB_FIELD_PAD, pad);
 }
